Added new methods to Operator class. Some changes made in Company class and Employee interface.

+++ src/Company.java
@@ -22,7 +22,7 @@ public class Company {
         staffList.remove(employee);
     }
 
-    protected int getIncome() {
+    protected double getIncome() {
         return income;
     }
